Make it obvious where to get the keys for hosted accounts
Sentry.io does not have the most straightforward settings and happened to me several times that I was trying to generate something called "deploy token" at the project level (which is IMO reasonable assumption). And it did not work. Only to find later on that the docs here actually give me hint where to obtain the key.

Somehow I always ignored the comment. This does not make it fool-proof, but IMO slightly more visible. So hopefully more people will have a success on the first go?